recovery: Provide caching for sideload files
Create a cache of block data received via adb.  The cache size is set
to ensure that there is at least 400mb available for the installer.

When the cache is large enough to hold the entire file, each block is
read via adb at most once.

When the cache is not large enough to hold the entire file, the cache
will need to be pruned.  Because files tend to be read sequentially
during install, the pruning algorithm attempts to discard blocks that
are behind the current file position.

Change-Id: Id8fc7fa5b38f1d80461eb576b1a1b5d53453cfc1
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>
Signed-off-by: Anush02198 <Anush.4376@gmail.com>